app-async/src/requirements.txt
Flask[async]
flask
Flask-SQLAlchemy
Flask-JWT-Extended
requests
postgrest
psycopg2-binary
asyncpg